Split integration tests by type of alert
This makes it easier to distinguish the test messages generated by
different types of alerts. It will also make it easier to add logic
that's specific to the ECR scan, since it will soon require an extra
API call.
